Mod11

+++ todo_app/app.py
@@ -2,11 +2,12 @@ from todo_app.view_model import ViewModel
 from .data.item import Item
 from .flask_config import Config
 import os
-from flask import Flask, redirect, render_template, request
+from flask import Flask, redirect, render_template, request, url_for,session
 import pymongo
 from bson import ObjectId # For ObjectId to work
-
-
+from todo_app.oauth import blueprint
+from flask_dance.contrib.github import github
+from werkzeug.middleware.proxy_fix import ProxyFix
 
 def create_app():
     dbconnection = os.getenv("MONGODB_CONN")
@@ -15,8 +16,9 @@ def create_app():
     collection = db.todolist
     
     app = Flask(__name__)
+    app.wsgi_app = ProxyFix(app.wsgi_app)
     app.config.from_object(Config())
-    
+    app.register_blueprint(blueprint, url_prefix="/login")
     
     def todolists():
         dicttodo=collection.find({"Status" : "To Do"})
@@ -41,14 +43,20 @@ def create_app():
     
     @app.route('/')
     def index():
+        if not github.authorized:
+            return redirect(url_for("github.login"))
         todo_items = todolists()
         doing_items = doinglists()
         done_item = donelists()
         item_view_model = ViewModel(todo_items, doing_items, done_item)
-        return render_template('index.html', view_model = item_view_model)
+        username=session["username"]
+        role = session["role"]
+        return render_template('index.html', view_model = item_view_model,username=username,role=role)
     
     @app.route('/additem', methods = ["POST"])
     def add_new_item():
+        if not github.authorized:
+            return redirect(url_for("github.login"))
         name=request.values.get("name")
         desc=request.values.get("desc")
         collection.insert_one({ "name":name, "desc":desc, "Status": "To Do"})
@@ -56,6 +64,8 @@ def create_app():
 
     @app.route('/movedoing', methods = ["POST"])
     def movedoing():
+        if not github.authorized:
+            return redirect(url_for("github.login"))
         id=request.values.get("_id")
         collection.find({"_id":ObjectId(id)})
         collection.update_one({"_id":ObjectId(id)}, {"$set": {"Status":"In Progress"}})
@@ -64,6 +74,8 @@ def create_app():
 
     @app.route('/movedone', methods = ["POST"])
     def movedone():
+        if not github.authorized:
+            return redirect(url_for("github.login"))
         id=request.values.get("_id")
         collection.find({"_id":ObjectId(id)})
         collection.update_one({"_id":ObjectId(id)}, {"$set": {"Status":"Completed"}})
@@ -71,6 +83,8 @@ def create_app():
 
     @app.route('/delete', methods = ["POST"])
     def remove():
+        if not github.authorized:
+            return redirect(url_for("github.login"))
         id=request.values.get("_id")
         collection.find({"_id":ObjectId(id)})
         collection.delete_one({"_id":ObjectId(id)})
